update source code
Changed logic to inlcude a while loop that only runs until the input is < the numeral values in the tuple. If input = 5 and the numeral value is 1000  the code will move to the next numeral until reahcing 5/5, then it continues as normal until it has accounted for the remaining numeral values. Alternative implementation is to check if the remainder is 0.